Add rarity to InstantLendTable and LoansTable
