+++ src/api/core/types.ts
@@ -77,6 +77,22 @@ export interface FetchMarketOffersResponse {
   meta: PaginationMeta
 }
 
+export enum RarityTier {
+  Common = 'common',
+  Uncommon = 'uncommon',
+  Rare = 'rare',
+  Epic = 'epic',
+  Legendary = 'legendary',
+  Mythic = 'mythic',
+}
+
+const RaritySchema = z.object({
+  tier: z.nativeEnum(RarityTier),
+  rank: z.number(),
+})
+
+export type Rarity = z.infer<typeof RaritySchema>
+
 export const NFTSchema = z.object({
   mint: z.string(),
   meta: z.object({
@@ -98,6 +114,7 @@ export const NFTSchema = z.object({
     })
     .optional(),
   collectionFloor: z.number(),
+  rarity: RaritySchema.optional(),
 })
 
 const BondTradeTransactionSchema = z.object({
@@ -174,20 +191,6 @@ export interface WalletLoansAndOffersResponse {
   meta: PaginationMeta
 }
 
-export enum RarityTier {
-  Common = 'common',
-  Uncommon = 'uncommon',
-  Rare = 'rare',
-  Epic = 'epic',
-  Legendary = 'legendary',
-  Mythic = 'mythic',
-}
-
-const RaritySchema = z.object({
-  tier: z.nativeEnum(RarityTier),
-  rank: z.number(),
-})
-
 const BorrowNftSchema = z.object({
   mint: z.string(),
   loan: z.object({
@@ -196,7 +199,7 @@ const BorrowNftSchema = z.object({
     marketApr: z.number(),
     banxStake: z.string().optional(),
   }),
-  nft: NFTSchema.extend({ rarity: RaritySchema.optional() }),
+  nft: NFTSchema,
 })
 
 export type BorrowNft = z.infer<typeof BorrowNftSchema>
