+++ src/components/TableComponents/TableCells.module.less
@@ -161,3 +161,10 @@
   display: flex;
   align-items: center;
 }
+
+.rarityText {
+  font: var(--body-text-sm);
+  border-radius: 100px;
+  padding: 2px 8px;
+  justify-content: flex-end;
+}
