+++ src/pages/BorrowPage/RequestLoansContent/components/RequestLoansTable/RequestLoansTable.module.less
@@ -13,13 +13,6 @@
   width: 96px;
 }
 
-.rarityText {
-  font: var(--body-text-sm);
-  border-radius: 100px;
-  padding: 2px 8px;
-  justify-content: flex-end;
-}
-
 .cellText {
   font: var(--body-text-sm);
 }
