+++ src/pages/LendPage/InstantLendContent/InstantLendTable/InstantLendTable.module.less
@@ -8,7 +8,7 @@
 }
 
 .refinanceTable {
-  min-width: 972px;
+  min-width: 1080px;
 
   tr > td,
   tr > th {
@@ -29,7 +29,7 @@
   }
 
   @media (max-width: 640px) {
-    min-width: 900px;
+    min-width: 990px;
 
     tr > td,
     tr > th {
