+++ src/pages/OffersPage/components/ActiveTabContent/components/LoansTable/LoansTable.module.less
@@ -20,7 +20,7 @@
   }
 
   @media (max-width: 640px) {
-    min-width: 792px;
+    min-width: 888px;
 
     tr > td,
     tr > th {
@@ -181,4 +181,8 @@
   align-items: center;
   column-gap: 4px;
   font: var(--body-text-md);
+}
+
+.snowflakeIcon path:first-child {
+  fill: var(--content-primary);
 }
